Add files via upload
5. Days of the Month:

By stimulating the scenario provided, using a switch statement to output the month and how many days containing base on user input number. Associating the month for corresponding number cases. 
At the end of case 12- December continue with Default Case to generate an output error if user input exceed number is  >12 month or <1, or user input a character.
 Also providing comments for each line of the code.