Add files via upload
8. Simple Search
To complete the optional requirements, program will take an user input name.
Defining a string array and associate 6 names to it, and using a linear search through the list an output will be generated with the appropriate message if the name was found or not on the list based on user input. Also by declaring a <cctype> and convert string tolowercase dave or DavE to be recognized and found on the list.